added subdirectory for cpp code
